Address warnings that prevented compilation with recent versions of clang. Take the opportunity to make a few improvements: use C++11 atomic variables instead of OS-specific calls, remove sanity checks that appear to be redundant with boost asserts, fix compilation without boost asserts to improve performance, allow float equality checks that are in fact OK to perform in a naive way (there are probably more comparisons that could be similarly streamlined), rename a function to better reflect what it does, and remove another that is hardly used.

+++ C++_plugins/plugin_dummy.cpp
@@ -43,8 +43,8 @@ void dummyPlugin(ImageIO* inputImage, ImageIO* outputImage,
 	float newPixelValue = pixelValue * 2;
 
 	// print to standard output
-	log(cb, 4, "old pixel value = %f, new pixel value = %f", pixelValue,
-			newPixelValue);
+	log(cb, 4, "old pixel value = %f, new pixel value = %f", double(pixelValue),
+			double(newPixelValue));
 
 	// replace old pixel value in I(x,y,z)
 	I(x, y, z) = newPixelValue;
